paulie

+++ drl_modules/obs_spaces.py
@@ -1,6 +1,6 @@
-import gym
+import gymnasium as gym
 import numpy as np
-from gym import spaces
+from gymnasium import spaces
 import pandas as pd
 import drl_modules.data_extract as de
 
@@ -33,7 +33,7 @@ def get_obs_space(df: pd.DataFrame, box_length: int = 10):
         'low': np.array([df["low"].min() - std_low] * box_length, dtype=np.float32),
         'close': np.array([df["close"].min() - std_close] * box_length, dtype=np.float32),
         'volume': np.array([0] * box_length, dtype=np.float32),
-        'position': np.array([0], dtype=np.float32)  # Assuming position type is a single value
+        'position': np.array([0], dtype=np.int32)  # Assuming position type is a single value
     }
     
     high_dict = {
@@ -42,7 +42,7 @@ def get_obs_space(df: pd.DataFrame, box_length: int = 10):
         'low': np.array([df["low"].max() + std_low] * box_length, dtype=np.float32),
         'close': np.array([df["close"].max() + std_close] * box_length, dtype=np.float32),
         'volume': np.array([df["tick_volume"].max() + std_vol] * box_length, dtype=np.float32),
-        'position': np.array([2], dtype=np.float32)  # Assuming position type is a single value
+        'position': np.array([2], dtype=np.int32)  # Assuming position type is a single value
     }
     
     return spaces.Dict({
